// apps/web/src/lib/api/generated/model/itemOut.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * BrokerOS API
 * OpenAPI spec version: 1.0.0
 */

export interface ItemOut {
  active: boolean;
  created_at: string;
  currency: string;
  description: string | null;
  id: string;
  name: string;
  tax_code: string;
  unit: string | null;
  /** @pattern ^-?\d{1,16}(\.\d{1,4})?$ */
  unit_price: string;
  version: number;
}